+++ zerver/tests/test_home.py
@@ -194,6 +194,8 @@ class HomeTest(ZulipTestCase):
         "server_generation",
         "server_inline_image_preview",
         "server_inline_url_embed_preview",
+        "server_max_deactivated_realm_deletion_days",
+        "server_min_deactivated_realm_deletion_days",
         "server_name_changes_disabled",
         "server_needs_upgrade",
         "server_presence_offline_threshold_seconds",